feature: Add better cbar behaviour to plotting
